CommunityTableViewCell 파일 분리 및 minimal bug fix
