+++ OffOff_iOS/OffOff_iOS/Cells/CommunityTableViewCell.swift
@@ -1,0 +1,46 @@
+//
+//  CommunityTableViewCell.swift
+//  OffOff_iOS
+//
+//  Created by Lee Nam Jun on 2021/07/18.
+//
+
+import UIKit
+
+class CommunityTableViewCell: UITableViewCell {
+    var titleImage: UIImageView!
+    var titleLabel: UILabel!
+    static let identifier = "CommunityTableViewCell"
+        
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupCell()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupCell() {
+        titleImage = UIImageView(image: UIImage(systemName: "cross")).then { $0.tintColor = .mainColor }
+        titleLabel = UILabel().then {
+            $0.text = "자유게시판"
+            $0.textAlignment = .right
+        }
+        
+        contentView.addSubview(titleImage)
+        contentView.addSubview(titleLabel)
+        
+        titleLabel.snp.makeConstraints {
+            $0.right.equalToSuperview().inset(12)
+            $0.top.bottom.equalToSuperview().inset(8)
+        }
+        
+        titleImage.snp.makeConstraints {
+            $0.left.equalToSuperview().inset(12)
+            $0.centerY.equalToSuperview()
+            $0.height.equalTo(titleLabel.snp.height)
+            $0.width.equalTo(titleImage.snp.height)
+        }
+    }
+}
